Split partial_json benchmarks (#18)
* Split partial_json benchmark

* Format partial_json benches
