+++ .github/workflows/benchmarks.yml
@@ -13,7 +13,9 @@ jobs:
       matrix:
         bench:
           - competitive_benchmarks
-          - partial_json
+          - partial_json_strategies
+          - partial_json_big
+          - partial_json_incremental
           - streaming_parser
     steps:
       - uses: actions/checkout@v4
