+++ crates/jsonmodem/Cargo.toml
@@ -45,7 +45,15 @@ perf = "deny"
 suspicious = "deny"
 
 [[bench]]
-name = "partial_json"
+name = "partial_json_strategies"
+harness = false
+
+[[bench]]
+name = "partial_json_big"
+harness = false
+
+[[bench]]
+name = "partial_json_incremental"
 harness = false
 
 [[bench]]
